+++ core/defaults/src/main/java/org/eclipse/dataspaceconnector/core/defaults/negotiationstore/InMemoryContractNegotiationStore.java
@@ -36,7 +36,6 @@ import java.util.concurrent.locks.ReentrantReadWriteLock;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static java.lang.String.format;
 import static java.util.stream.Collectors.toList;
 
 /**
@@ -122,14 +121,6 @@ public class InMemoryContractNegotiationStore implements ContractNegotiationStor
     @Override
     public @NotNull Stream<ContractAgreement> queryAgreements(QuerySpec querySpec) {
         return lockManager.readLock(() -> agreementQueryResolver.query(getAgreements(), querySpec));
-    }
-
-    @Override
-    public Stream<ContractNegotiation> getNegotiationsWithAgreementOnAsset(String assetId) {
-        var filter = format("contractAgreement.assetId = %s", assetId);
-        var query = QuerySpec.Builder.newInstance().filter(filter).build();
-
-        return queryNegotiations(query);
     }
 
     @Override